+++ PXCTestKit/Command/RunTestsCommand.swift
@@ -24,39 +24,36 @@ final class RunTestsCommand {
         let testsToRun: Set<String>
         let simulators: [FBSimulatorConfiguration]
         let timeout: Double
+
+        func logFileURL() -> URL {
+            return output.appendingPathComponent("simulator.log")
+        }
     }
 
     private let configuration: Configuration
-    private let logFileURL: URL
-
-    private var application: FBApplicationDescriptor!
-    private var testRun: FBXCTestRun!
 
     init(configuration: Configuration) {
         self.configuration = configuration
-        self.logFileURL = configuration.output.appendingPathComponent("simulator.log")
     }
 
     func run() throws {
-        try XCTestBootstrapFrameworkLoader.loadPrivateFrameworks(nil)
+        let testRun = try FBXCTestRun.withTestRunFile(atPath: configuration.testRun.path).build()
 
-        try resetOutput()
+        try resetOutput(testRun: testRun)
 
-        testRun = try FBXCTestRun.withTestRunFile(atPath: configuration.testRun.path).build()
-        application = try FBApplicationDescriptor.application(withPath: testRun.testHostPath)
-
-        let logFileHandle = try FileHandle(forWritingTo: logFileURL)
-        let logger = FBControlCoreLogger.aslLoggerWriting(toFileDescriptor: logFileHandle.fileDescriptor, withDebugLogging: false)
-        let simulatorControlConfiguration = FBSimulatorControlConfiguration(deviceSetPath: configuration.deviceSet.path, options: [])
-        let control = try FBSimulatorControl.withConfiguration(simulatorControlConfiguration, logger: logger)
+        let logFileHandle = try FileHandle(forWritingTo: configuration.logFileURL())
+        let control = try FBSimulatorControl.withConfiguration(
+            FBSimulatorControlConfiguration(deviceSetPath: configuration.deviceSet.path, options: []),
+            logger: FBControlCoreLogger.aslLoggerWriting(toFileDescriptor: logFileHandle.fileDescriptor, withDebugLogging: false)
+        )
 
         let simulators = try configuration.simulators.map {
             try control.pool.allocateSimulator(with: $0, options: [.create, .reuse])
         }
 
         try boot(simulators: simulators)
-        try test(simulators: simulators)
-        try extractDiagnostics(simulators: simulators)
+        try test(simulators: simulators, testRun: testRun)
+        try extractDiagnostics(simulators: simulators, testRun: testRun)
 
         ConsoleReporter.writeSummary()
 
@@ -67,7 +64,7 @@ final class RunTestsCommand {
 
     // MARK: - Private
 
-    private func resetOutput() throws {
+    private func resetOutput(testRun: FBXCTestRun) throws {
         let fileManager = FileManager.default
 
         if fileManager.fileExists(atPath: configuration.output.path) {
@@ -75,10 +72,13 @@ final class RunTestsCommand {
         }
 
         try fileManager.createDirectory(at: configuration.output, withIntermediateDirectories: true, attributes: nil)
-        fileManager.createFile(atPath: logFileURL.path, contents: nil, attributes: nil)
+        fileManager.createFile(atPath: configuration.logFileURL().path, contents: nil, attributes: nil)
 
-        for simulatorConfiguration in configuration.simulators {
-            try fileManager.createDirectory(at: outputURL(simulatorConfiguration: simulatorConfiguration), withIntermediateDirectories: true, attributes: nil)
+        for target in testRun.targets {
+            for simulatorConfiguration in configuration.simulators {
+                let url = outputURL(for: simulatorConfiguration, target: target)
+                try fileManager.createDirectory(at: url, withIntermediateDirectories: true, attributes: nil)
+            }
         }
     }
 
@@ -100,59 +100,61 @@ final class RunTestsCommand {
         }
     }
 
-    private func test(simulators: [FBSimulator]) throws {
-        var environment = testRun.environment
-        // Make Xcode output less noisy
-        environment["OS_ACTIVITY_MODE"] = "disable"
+    private func test(simulators: [FBSimulator], testRun: FBXCTestRun) throws {
+        for target in testRun.targets {
+            var testEnvironment = target.testLaunchConfiguration.testEnvironment ?? [:]
+            testEnvironment["OS_ACTIVITY_MODE"] = "disable"
+            let testsToRun = target.testLaunchConfiguration.testsToRun.union(configuration.testsToRun)
+            let testLaunchConfigurartion = target.testLaunchConfiguration
+                .withTestsToRun(testsToRun)
+                .withTestEnvironment(testEnvironment)
 
-        let applicationLaunchConfiguration = FBApplicationLaunchConfiguration(
-            application: application,
-            arguments: testRun.arguments,
-            environment: environment,
-            options: []
-        )
+            for simulator in simulators {
+                let simulatorIdentifier = "\(simulator.configuration!.deviceName) \(simulator.configuration!.osVersionString)"
+                let consoleReporter = ConsoleReporter(simulatorIdentifier: simulatorIdentifier, testTargetName: target.name)
+                let junitReportURL = outputURL(for: simulator.configuration!, target: target)
+                let junitReporter = FBTestManagerTestReporterJUnit.withOutputFileURL(junitReportURL.appendingPathComponent("junit.xml"))
 
-        let testLaunchConfiguration = FBTestLaunchConfiguration(testBundlePath: testRun.testBundlePath)
-            .withApplicationLaunchConfiguration(applicationLaunchConfiguration)
-            .withTimeout(configuration.timeout)
-            .withTestsToSkip(testRun.testsToSkip)
-            .withTestsToRun(testRun.testsToRun.union(configuration.testsToRun))
+                for application in target.applications {
+                    try simulator.interact
+                        .installApplication(application)
+                        .perform()
+                }
 
-        for simulator in simulators {
-            let simulatorConfiguration = simulator.configuration!
-            let consoleReporter = ConsoleReporter(simulatorIdentifier: identifier(simulatorConfiguration: simulatorConfiguration))
-            let junitReporter = FBTestManagerTestReporterJUnit.withOutputFileURL(outputURL(simulatorConfiguration: simulatorConfiguration).appendingPathComponent("junit.xml"))
-            try simulator.interact
-                .installApplication(application)
-                .startTest(
-                    with: testLaunchConfiguration,
-                    reporter: FBTestManagerTestReporterComposite.withTestReporters([consoleReporter, junitReporter, SummaryReporter()])
-                )
-                .perform()
-        }
+                try simulator.interact
+                    .startTest(
+                        with: testLaunchConfigurartion,
+                        reporter: FBTestManagerTestReporterComposite.withTestReporters([consoleReporter, junitReporter, SummaryReporter()])
+                    )
+                    .perform()
+            }
 
-        for simulator in simulators {
-            try simulator.interact
-                .waitUntilAllTestRunnersHaveFinishedTesting(withTimeout: configuration.timeout)
-                .perform()
-        }
-    }
-
-    private func extractDiagnostics(simulators: [FBSimulator]) throws {
-        for simulator in simulators {
-            guard let diagnostics = simulator.diagnostics.launchedProcessLogs().first(where: { $0.0.processName == application.name })?.value else { continue }
-            guard let logFilePath = diagnostics.asPath else { return }
-            let destinationPath = outputURL(simulatorConfiguration: simulator.configuration!).appendingPathComponent("\(application.name).log").path
-            try FileManager.default.copyItem(atPath: logFilePath, toPath: destinationPath)
+            for simulator in simulators {
+                try simulator.interact
+                    .waitUntilAllTestRunnersHaveFinishedTesting(withTimeout: configuration.timeout)
+                    .perform()
+            }
         }
     }
 
-    private func outputURL(simulatorConfiguration: FBSimulatorConfiguration) -> URL {
-        return configuration.output.appendingPathComponent(identifier(simulatorConfiguration: simulatorConfiguration))
+    private func extractDiagnostics(simulators: [FBSimulator], testRun: FBXCTestRun) throws {
+        for simulator in simulators {
+            for target in testRun.targets {
+                for application in target.applications {
+                    guard let diagnostics = simulator.diagnostics.launchedProcessLogs().first(where: { $0.0.processName == application.name })?.value else { continue }
+                    guard let logFilePath = diagnostics.asPath else { return }
+                    let destinationPath = outputURL(for: simulator.configuration!, target: target).appendingPathComponent("\(application.name).log").path
+                    try FileManager.default.copyItem(atPath: logFilePath, toPath: destinationPath)
+                }
+            }
+        }
     }
 
-    private func identifier(simulatorConfiguration: FBSimulatorConfiguration) -> String {
-        return "\(simulatorConfiguration.deviceName) - \(simulatorConfiguration.osVersionString)"
+    private func outputURL(for simulatorConfiguration: FBSimulatorConfiguration, target: FBXCTestRunTarget) -> URL {
+        return configuration.output
+            .appendingPathComponent(target.name)
+            .appendingPathComponent(simulatorConfiguration.osVersionString)
+            .appendingPathComponent(simulatorConfiguration.deviceName)
     }
 
 }